dgo-gen: sync the change of API from Go

+++ dgo-gen/internal/generator/gogen/gogen.go
@@ -279,10 +279,10 @@ func buildFunction_DgoLoad(etype *exported.Type, term ir.Term, g *Group, looper 
 	case *ir.PinToken:
 		g.BlockFunc(func(g *Group) {
 			g.Var().Id("version").Uint16()
-			g.Var().Id("lid").Uint16()
+			g.Var().Id("lid").Uint8()
 			g.Var().Id("data").Uintptr()
 			loadIntoInt(g, arrIndex(), Id("version"), Uint16())
-			loadIntoInt(g, arrIndex(), Id("lid"), Uint16())
+			loadIntoInt(g, arrIndex(), Id("lid"), Uint8())
 			loadIntoInt(g, arrIndex(), Id("data"), Uintptr())
 			g.Op("*o").Op("=").
 				Add(Parens(typeNameOf(etype, t))).
@@ -805,7 +805,7 @@ func (d *Generator) buildStub(dstDir string, pkgName string) {
 		Comment("//go:linkname pin_untypedTokenFromRaw github.com/hsfzxjy/dgo/go/pin.untypedTokenFromRaw").
 		Line().
 		Func().Id("pin_untypedTokenFromRaw").
-		Params(Id("version").Uint16(), Id("lid").Uint16(), Id("data").Uintptr()).
+		Params(Id("version").Uint16(), Id("lid").Uint8(), Id("data").Uintptr()).
 		Id("pin_untypedToken").
 		Line().
 		Line().
@@ -819,7 +819,7 @@ func (d *Generator) buildStub(dstDir string, pkgName string) {
 		Line().
 		Func().Id("pin_untypedTokenExtract").
 		Params(Id("token").Add(untypedPinToken())).
-		Parens(List(Id("version").Uint16(), Id("lid").Uint16(), Id("data").Uintptr())).
+		Parens(List(Id("version").Uint16(), Id("lid").Uint8(), Id("data").Uintptr())).
 		Line()
 }
